:bug: Fix gh directory non-initialization issue
For a non-initialized GH folder, preivously the function was trying to
initiate a repo only when the directory is not found. So, in the case
that the directory already exists, but not repo is initiated, this
causes a problem. To mitigate it, now `git init` is called always, and
if failed, the exception is printed.